Updated Credentialler and it's tests to reflect the builder changes
